Use the new API of Labeled_mesh_domain_3 for the demo

+++ Polyhedron/demo/Polyhedron/Plugins/Mesh_3/Polyhedron_demo_mesh_3_labeled_mesh_domain_3.h
@@ -37,13 +37,13 @@ namespace CGAL {
 
 struct Compare_to_isovalue {
   typedef bool result_type;
-  float iso_value;
+  double iso_value;
   bool less;
 
-  Compare_to_isovalue(float iso_value, bool less)
+  Compare_to_isovalue(double iso_value, bool less)
     : iso_value(iso_value), less(less) {}
 
-  bool operator()(float x) const {
+  bool operator()(double x) const {
     return (x < iso_value) == less;
   }
 };
